feat: Add ArticleDetail screen and navigation
- Introduces `ArticleDetailScreen` and `ArticleDetailNavigation` for displaying detailed article content.
- Implements `ArticleDetailViewModel` to fetch and manage the state of an article's details.
- Adds `ArticleDetailUiState` to represent the UI state of the ArticleDetail screen.
- Creates `ArticleDetailRoute` Composable function for the `ArticleDetailScreen`.
- Adds Koin configuration for `ArticleDetailViewModel`.
- Updates `BlogNavHost` to include the `articleDetailScreen` in the navigation.
- Modifies `ArticleApi` to use `Long` for article IDs.
- Updates `ArticleRepository` to use `Long` for article IDs.
- Updates `ArticlesScreen` to navigate to `ArticleDetail` screen.
- Removes `DestinationUtils.kt`.
- Creates new `DestinationUtils.kt`.
- Updates `Window.wasmJs.kt` to use new navigation implementation.
- Update `Destinations.kt` to handle route pattern and parsing.

+++ frontend/src/commonMain/kotlin/me/matsumo/blog/core/repository/ArticleRepository.kt
@@ -18,7 +18,7 @@ class ArticleRepository(
         articleApi.getArticles().let(articleMapper::map)
     }
 
-    suspend fun getArticleDetail(id: Int): ArticleDetail = withContext(ioDispatcher) {
+    suspend fun getArticleDetail(id: Long): ArticleDetail = withContext(ioDispatcher) {
         articleApi.getArticleDetail(id).let(articleMapper::map)
     }
 }
